Moved script templates to a templates folder and added filters to be able to render the script tag with the nonce
